Allow extra counts to be skipped when loading contact grid.
We already have the lead count, and that extra lead count costs us many
seconds to query.  In addition supporting a dynamic orderBy allows us
to avoid a tmp table file sort on millions of rows, but we’ll get
effectively the same intended result, being the most recent leads added
to the campaign in 3ms vs 89s.